remove catch-all step
the catch-all step was too smart for its own good. it was beginning to cause trouble when, for example, a step with an argument was matching to the catch-all step instead of the specific step as I expected. removing it also allowed me to remove all the --guess parameters to Cucumber.