fix "%" bug
this was a bug that happened when you had a % character in any string that was sent to the server.
it happened because I'd gone through unnecessary hoops (converting data to json), and I'd gotten confused by the encodings.
I removed json and just used POST parameters instead - this fixed it.
no need to add a permanent test, I think - this was a very specific problem, and by removing json, it should be gone forever